Consolidate duplicated entity lifecycle boilerplate into a mixin
update_callback/async_update/async_will_remove_from_hass were hand-rolled
identically in all four entity classes (EmeraldWaterHeater and the three
energy sensors) -- the only thing that ever differed between them was what
update() itself does. Extracted CallbackDrivenEntityMixin in helpers.py;
each class now just lists it first in its bases and implements update().

Verified via MRO inspection (super() chains correctly through the mixin to
the real HA Entity base) and an instantiation/callback smoke test. Net -55
lines.
